requirements.txt
Flask==3.0.0
line_bot_sdk==3.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1